+++ src/main/java/com/example/my_books_backend/entity/BookChapterId.java
@@ -16,5 +16,5 @@ public class BookChapterId implements Serializable {
     private String bookId;
 
     @Column(name = "chapter_number")
-    private Integer chapterNumber;
+    private Long chapterNumber;
 }
